This function scrapes the openFDA API for food product recall enforcement reports based on user inputs of date and time

This function scrapes the openFDA API for food product recall enforcement reports based on user inputs of date and time

Usage

recall_date(
  api_key,
  center_classification_date = NULL,
  limit = NULL,
  product_description = NULL,
  recall_initiation_date = NULL,
  recalling_firm = NULL,
  report_date = NULL,
  search_mode = NULL,
  status = NULL,
  termination_date = NULL
)

Arguments

api_key

Your free api key from FDA website

center_classification_date

A way the FDA classifies a date

limit

The number of rows to return for that query

product_description

Description of product

recall_initiation_date

Date for which recall was initiated

recalling_firm

The company recalling the product

report_date

The date the FDA issued the enforcement report for the product recall

search_mode

This gives the user flexibility to search for exact matches of inputs or any combination of inputs

status

The status of the recall

termination_date

The date the recall was terminated

Value

A data frame with the returned results of the users query to the API

Examples